requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.0.8
django-image-optimizer==1.0.1
django-mathfilters==1.0.0
django-robots==4.0
dnspython==1.16.0
email-validator==1.1.1
filelock==3.6.0
gunicorn==20.1.0
idna==2.10
Pillow==7.2.0

py3-validate-email==0.2.9
python-resize-image==1.1.20
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tinify==1.5.2
urllib3==1.26.9
